Create Scikit-image: image processing in Python
scikit-image is an image processing Python package that works with numpy arrays which is a collection of algorithms for image processing. Let’s discusses how to deal with images into set of information and its some application in real world.
Important features of scikit-image :
Simple and efficient tools for image processing and computer vision techniques.
Accessible to everybody and reusable in various contexts.
Built on the top of NumPy, SciPy, and matplotlib.
Open source, commercially usable – BSD license.