Create Matlab thermal temperature plot (Arduino)
Creating MATLAB Graphical User Interface for Plotting Graph
First we have to build GUI (Graphical User Interface) for Plot a Graph with the temperature data. To launch the GUI, type the below command in the command window
guide
A popup window will open, then select new blank GUI .
Now we have to choose one push button, two axes and one text box for MATLAB graphical interface. Push button will be used for start the temperature sensing, two axes for plotting the graph and Text box to show the current value of temperature.
To resize or to change the shape of the button, Axes or edit text button, just click on it and you will be able to drag the corners of the button. By double-clicking on any of these you will be able to change the color, string and tag of that particular button.
You can customize the buttons as per your choice. Now when you save this, a code is generated in the Editor window of MATLAB. To make your Arduino to perform any particular task related to your project, you always have to edit this generated code.
MATLAB Code for Plotting the Graph
Complete MATLAB code for Designing a Thermometer using LM35 and Arduino, is given at the end of this project. Further we are including the GUI file (.fig) and code file(.m) here for download, using which you can customize the buttons or Axes size as per your requirement. We have edited the generated code as explained below.
Copy and paste the below code on line no. 74 to make sure that the Arduino is talking with MATLAB every time you run the m-file.
Material Required
MATLAB installed Laptop (Preference: R2016a or above versions)
Arduino UNO
LM35 – Temperature Sensor
Connecting Wires
Breadboard
CIRCUIT DIAGRAM: